Enables moving windows on a second monitor
Allows the user to use the Shift + Win Key + Number to be able to move windows to a specific desktop, if those windows are on a secondary monitor. Previously, this feature only worked for windows which were on the primary monitor. This does not support more monitors than 2 as I do not have the ability to test more than 2 monitors.